Fix error by mention by reviewer
