+++ src/components/Calculator.js
@@ -1,7 +1,97 @@
-
+import { useState, useEffect } from 'react';
+import NumberFormat from 'react-number-format';
 
 const Calculator = () => {
-  
+  const [preState, setPreState] = useState('');
+  const [curState, setCurState] = useState('');
+  const [input, setInput] = useState('0');
+  const [operator, setOperator] = useState(null);
+  const [total, setTotal] = useState(false);
+
+  const inputNum = (e) => {
+    if (curState.includes('.') && e.target.innerText === '.') return;
+
+    if (total) {
+      setPreState('');
+    }
+    if (curState) {
+      setCurState((pre) => pre + e.target.innerText);
+    } else {
+      setCurState(e.target.innerText);
+      setTotal(false);
+    }
+  };
+
+  useEffect(() => {
+    setInput(curState);
+  }, [curState]);
+
+  useEffect(() => {
+    setInput('0');
+  }, []);
+
+  // equals
+
+  const equals = (e) => {
+    if (e?.target.innerText === '=') {
+      setTotal(true);
+    }
+    let cal;
+    switch (operator) {
+      case '÷':
+        cal = String(parseFloat(preState) / parseFloat(curState));
+        break;
+
+      case '+':
+        cal = String(parseFloat(preState) + parseFloat(curState));
+        break;
+      case 'X':
+        cal = String(parseFloat(preState) * parseFloat(curState));
+        break;
+      case '-':
+        cal = String(parseFloat(preState) - parseFloat(curState));
+        break;
+      default:
+        return;
+    }
+    setInput('');
+    setPreState(cal);
+    setCurState('');
+  };
+
+  const operatorType = (e) => {
+    setTotal(false);
+    setOperator(e.target.innerText);
+    if (curState === '') return;
+    if (preState !== '') {
+      equals();
+    } else {
+      setPreState(curState);
+      setCurState('');
+    }
+  };
+
+  const minusPlus = () => {
+    if (curState.charAt(0) === '-') {
+      setCurState(curState.substring(1));
+    } else {
+      setCurState(`-${curState}`);
+    }
+  };
+
+  const percent = () => {
+    if (preState) {
+      setCurState(String((parseFloat(curState) / 100) * preState));
+    } else {
+      setCurState(String(parseFloat(curState) / 100));
+    }
+  };
+
+  const reset = () => {
+    setPreState('');
+    setCurState('');
+    setInput('0');
+  };
 
   return (
     <div className="container">
@@ -24,7 +114,7 @@ const Calculator = () => {
         <button type="button" className="btn light-gray" onClick={minusPlus}>
           +/-
         </button>
-        <button type="button" className="btn light-gray" onClick={percent}>
+        <button type="button" className="btn light-gray row" onClick={percent}>
           %
         </button>
         <button type="button" className="btn orange" onClick={operatorType}>
@@ -36,7 +126,7 @@ const Calculator = () => {
         <button type="button" className="btn" onClick={inputNum}>
           8
         </button>
-        <button type="button" className="btn " onClick={inputNum}>
+        <button type="button" className="btn row" onClick={inputNum}>
           9
         </button>
         <button type="button" className="btn orange" onClick={operatorType}>
@@ -48,7 +138,7 @@ const Calculator = () => {
         <button type="button" className="btn" onClick={inputNum}>
           5
         </button>
-        <button type="button" className="btn " onClick={inputNum}>
+        <button type="button" className="btn row" onClick={inputNum}>
           6
         </button>
         <button type="button" className="btn orange" onClick={operatorType}>
@@ -60,7 +150,7 @@ const Calculator = () => {
         <button type="button" className="btn" onClick={inputNum}>
           2
         </button>
-        <button type="button" className="btn " onClick={inputNum}>
+        <button type="button" className="btn row" onClick={inputNum}>
           3
         </button>
         <button type="button" className="btn orange" onClick={operatorType}>
@@ -69,7 +159,7 @@ const Calculator = () => {
         <button type="button" className="btn zero" onClick={inputNum}>
           0
         </button>
-        <button type="button" className="btn " onClick={inputNum}>
+        <button type="button" className="btn row1" onClick={inputNum}>
           .
         </button>
         <button type="button" className="btn orange" onClick={equals}>
